requirements.txt
langgraph==0.2.28
langchain-anthropic==0.2.3
anthropic==0.39.0
instructor[anthropic]
python-dotenv==1.0.1
httpx==0.27.2
mcp==1.0.0
pydantic==2.11.9
mcp==1.0.0